--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -48,55 +48,56 @@ jobs:
       - name: Validate `package.json`s
         run: npm run pjson:check
 
-  test:
-    runs-on: ubuntu-latest
+test:
+    runs-on: ${{ matrix.os }}
 
     strategy:
-      matrix:
-        node:
-          - version: 23.x
-          - version: 22.x
-          # glob is not backported below 22.x
-      fail-fast: false # prevent a failure in other versions run cancelling others
+        matrix:
+            os: [ubuntu-latest, windows-latest]
+            node:
+                - version: 23.x
+                - version: 22.x
+                # glob is not backported below 22.x
+        fail-fast: false # prevent a failure in other versions run cancelling others
 
     steps:
-      - name: Harden Runner
-        uses: step-security/harden-runner@4d991eb9b905ef189e4c376166672c3f2f230481 # v2.11.0
-        with:
-          egress-policy: audit
+        - name: Harden Runner
+          uses: step-security/harden-runner@4d991eb9b905ef189e4c376166672c3f2f230481 # v2.11.0
+          with:
+                egress-policy: audit
 
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-      - name: Use Node.js ${{ matrix.node.version }}
-        uses: actions/setup-node@cdca7365b2dadb8aad0a33bc7601856ffabcc48e # v4.3.0
-        with:
-          node-version: ${{ matrix.node.version }}
-          cache: 'npm'
+        - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+        - name: Use Node.js ${{ matrix.node.version }}
+          uses: actions/setup-node@cdca7365b2dadb8aad0a33bc7601856ffabcc48e # v4.3.0
+          with:
+            node-version: ${{ matrix.node.version }}
+            cache: 'npm'
 
-      - name: npm clean install
-        run: npm ci
+        - name: npm clean install
+          run: npm ci
 
-      - run: node --run test
+        - run: node --run test
 
-      - name: Publish coverage to Coveralls ${{ matrix.test_number }}
-        uses: coverallsapp/github-action@648a8eb78e6d50909eff900e4ec85cab4524a45b # v2.3.6
-        with:
-          flag-name: run-${{ join(matrix.*, '-') }}
-          file: ./coverage.lcov
-          format: lcov
-          parallel: true
+        - name: Publish coverage to Coveralls ${{ matrix.os }}-${{ matrix.node.version }}
+          if: ${{ matrix.os == 'ubuntu-latest' }}
+          uses: coverallsapp/github-action@648a8eb78e6d50909eff900e4ec85cab4524a45b # v2.3.6
+          with:
+            flag-name: run-${{ join(matrix.*, '-') }}
+            file: ./coverage.lcov
+            format: lcov
+            parallel: true
 
-  finish:
+finish:
     needs: test
     if: ${{ success() }}
     runs-on: ubuntu-latest
     steps:
-      - name: Harden Runner
-        uses: step-security/harden-runner@4d991eb9b905ef189e4c376166672c3f2f230481 # v2.11.0
-        with:
-          egress-policy: audit
+        - name: Harden Runner
+          uses: step-security/harden-runner@4d991eb9b905ef189e4c376166672c3f2f230481 # v2.11.0
+          with:
+            egress-policy: audit
 
-      - name: Coveralls Finished
-        uses: coverallsapp/github-action@648a8eb78e6d50909eff900e4ec85cab4524a45b # v2.3.6
-        with:
-          parallel-finished: true
-          carryforward: "run-1,run-2"
+        - name: Coveralls Finished
+          uses: coverallsapp/github-action@648a8eb78e6d50909eff900e4ec85cab4524a45b # v2.3.6
+          with:
+            parallel-finished: true