--- a/packages/deno-npm-prefix/deno-npm-prefix.bench.mjs
+++ b/packages/deno-npm-prefix/deno-npm-prefix.bench.mjs
@@ -16,7 +16,7 @@ suite.add('--loader', { repeatSuite: 2 }, () => {
 		[
 			'--no-warnings',
 			'--loader',
-			fileURLToPath(import.meta.resolve('./deno-npm-prefix.mjs')),
+			import.meta.resolve('./deno-npm-prefix.mjs'),
 			e2eTest,
 		],
 		{
@@ -33,7 +33,7 @@ suite.add('--import (register)', { repeatSuite: 2 }, () => {
 		[
 			'--no-warnings',
 			'--import',
-			fileURLToPath(import.meta.resolve('./fixtures/register.mjs')),
+			import.meta.resolve('./fixtures/register.mjs'),
 			e2eTest,
 		],
 		{
@@ -51,7 +51,7 @@ if (process.version.startsWith('v23')) {
 			[
 				'--no-warnings',
 				'--import',
-				fileURLToPath(import.meta.resolve('./fixtures/register-hooks.mjs')),
+				import.meta.resolve('./fixtures/register-hooks.mjs'),
 				e2eTest,
 			],
 			{