--- a/packages/svelte/svelte.bench.mjs
+++ b/packages/svelte/svelte.bench.mjs
@@ -16,7 +16,7 @@ suite.add('--loader', { repeatSuite: 2 }, () => {
 		[
 			'--no-warnings',
 			'--loader',
-			fileURLToPath(import.meta.resolve('./svelte.mjs')),
+			import.meta.resolve('./svelte.mjs'),
 			e2eTest,
 		],
 		{
@@ -35,7 +35,7 @@ suite.add('--import (register)', { repeatSuite: 2 }, () => {
 		[
 			'--no-warnings',
 			'--import',
-			fileURLToPath(import.meta.resolve('./fixtures/register.mjs')),
+			import.meta.resolve('./fixtures/register.mjs'),
 			e2eTest,
 		],
 		{