Fixes #188

---

I'm not 100% familiar with this codebase, so maintainers have edit access, and I invite you to **please** commit any changes directly into this PR.