@AugustinMauroy 2025-04-04
- test suite fail
- `node --run bench -- --w="packages/alias"` to have benchmark and see if there are huge gap