@AugustinMauroy 2025-04-05
@JakobJingleheimer why not enable windows on ci ?