@AugustinMauroy 2025-04-05
> RE why not test windows in CI? I don't recall.

I don't think we've asked the same question before?