@AugustinMauroy 2025-04-05
On main:
```
✓ Found 9 benchmarks
ℹ Running packages/tsx/tsx.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | █████████████----------------- | 13.8 ops/sec | 18 samples
--import (register)                           | ██████████████████████████████ | 32.15 ops/sec | 18 samples
✓ Successfully ran packages/tsx/tsx.bench.mjs

ℹ Running packages/svelte/svelte.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ██████████████████████████████ | 5.03 ops/sec | 12 samples
--import (register)                           | ██████████████████████████████ | 4.97 ops/sec | 14 samples
✓ Successfully ran packages/svelte/svelte.bench.mjs

ℹ Running packages/parse-filename/parse-filename.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

getFilenameExt                                | ███████████████████████------- | 1,351,658 ops/sec | 20 samples
stripExtras                                   | ██████████████████████████████ | 1,748,739 ops/sec | 24 samples
getFilenameParts                              | ██████------------------------ | 348,002 ops/sec | 22 samples
✓ Successfully ran packages/parse-filename/parse-filename.bench.mjs

ℹ Running packages/mismatched-format/mismatched-format.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

CJS: require()                                | █████████████████████████████- | 22,311 ops/sec | 20 samples
CJS: module.exports                           | █████████████████████████████- | 22,722 ops/sec | 19 samples
ESM: createRequire                            | ██████████████████████████████ | 23,430 ops/sec | 20 samples
ESM: require() within a comment               | ██████████████████████████████ | 23,448 ops/sec | 19 samples
✓ Successfully ran packages/mismatched-format/mismatched-format.bench.mjs

ℹ Running packages/jsonc/jsonc.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ███████████████████----------- | 22.65 ops/sec | 21 samples
--import (register)                           | ██████████████████████████████ | 35.45 ops/sec | 17 samples
✓ Successfully ran packages/jsonc/jsonc.bench.mjs

ℹ Running packages/json5/json5.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ███████████████████----------- | 22.71 ops/sec | 19 samples
--import (register)                           | ██████████████████████████████ | 35.67 ops/sec | 19 samples
✓ Successfully ran packages/json5/json5.bench.mjs

ℹ Running packages/deno-npm-prefix/deno-npm-prefix.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ███████████████████████------- | 15.82 ops/sec | 16 samples
--import (register)                           | ██████████████████████-------- | 15.41 ops/sec | 17 samples
--import (registerHooks)                      | ██████████████████████████████ | 20.99 ops/sec | 20 samples
✓ Successfully ran packages/deno-npm-prefix/deno-npm-prefix.bench.mjs

ℹ Running packages/css-module/css-module.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ███████████████████----------- | 22.62 ops/sec | 20 samples
--import (register)                           | ██████████████████████████████ | 35.41 ops/sec | 17 samples
✓ Successfully ran packages/css-module/css-module.bench.mjs

ℹ Running packages/alias/alias.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ███████████████████----------- | 22.63 ops/sec | 22 samples
--import (register)                           | ██████████████████████████████ | 35.34 ops/sec | 15 samples
--import (registerHooks)                      | ██████████████████████████████ | 35.62 ops/sec | 20 samples
✓ Successfully ran packages/alias/alias.bench.mjs
```
This pr:
```
✓ Found 9 benchmarks
ℹ Running packages/tsx/tsx.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | █████████████----------------- | 14.35 ops/sec | 17 samples
--import (register)                           | ██████████████████████████████ | 33.22 ops/sec | 20 samples
✓ Successfully ran packages/tsx/tsx.bench.mjs

ℹ Running packages/svelte/svelte.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ████████████████████████████-- | 4.29 ops/sec | 13 samples
--import (register)                           | ██████████████████████████████ | 4.61 ops/sec | 12 samples
✓ Successfully ran packages/svelte/svelte.bench.mjs

ℹ Running packages/parse-filename/parse-filename.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

getFilenameExt                                | ██████████████████████████---- | 670,853 ops/sec | 19 samples
stripExtras                                   | ██████████████████████████████ | 766,983 ops/sec | 21 samples
getFilenameParts                              | ████████████------------------ | 309,555 ops/sec | 21 samples
✓ Successfully ran packages/parse-filename/parse-filename.bench.mjs

ℹ Running packages/mismatched-format/mismatched-format.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

CJS: require()                                | █████████████████████████████- | 21,349 ops/sec | 20 samples
CJS: module.exports                           | █████████████████████████████- | 21,339 ops/sec | 22 samples
ESM: createRequire                            | ██████████████████████████████ | 22,230 ops/sec | 20 samples
ESM: require() within a comment               | ██████████████████████████████ | 22,091 ops/sec | 19 samples
✓ Successfully ran packages/mismatched-format/mismatched-format.bench.mjs

ℹ Running packages/jsonc/jsonc.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ███████████████████----------- | 21.61 ops/sec | 16 samples
--import (register)                           | ██████████████████████████████ | 33.95 ops/sec | 20 samples
✓ Successfully ran packages/jsonc/jsonc.bench.mjs

ℹ Running packages/json5/json5.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ██████████████████------------ | 19.55 ops/sec | 21 samples
--import (register)                           | ██████████████████████████████ | 32.32 ops/sec | 20 samples
✓ Successfully ran packages/json5/json5.bench.mjs

ℹ Running packages/deno-npm-prefix/deno-npm-prefix.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ██████████████████████-------- | 14.85 ops/sec | 19 samples
--import (register)                           | █████████████████████--------- | 14.23 ops/sec | 18 samples
--import (registerHooks)                      | ██████████████████████████████ | 19.94 ops/sec | 20 samples
✓ Successfully ran packages/deno-npm-prefix/deno-npm-prefix.bench.mjs

ℹ Running packages/css-module/css-module.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ███████████████████----------- | 21.09 ops/sec | 21 samples
--import (register)                           | ██████████████████████████████ | 33.19 ops/sec | 19 samples
✓ Successfully ran packages/css-module/css-module.bench.mjs

ℹ Running packages/alias/alias.bench.mjs
Node.js version: v23.10.0
Platform: darwin arm64
CPU Cores: 8 vCPUs | 16.0GB Mem

--loader                                      | ██████████████████------------ | 20.17 ops/sec | 20 samples
--import (register)                           | ██████████████████████████████ | 33.68 ops/sec | 20 samples
--import (registerHooks)                      | ███████████████████████████--- | 30.64 ops/sec | 20 samples
✓ Successfully ran packages/alias/alias.bench.mjs
```